[DO NOT UPSTREAM. Strictly confidential] Motion sensor for nRF52 Desktop project.
This commit provides mouse motion sensor for nRF52 Desktop.

The commit title is lengthy on purpose to BREAK the compliance check
in NCS repository in order not to be merged, published nor upstreamed
before an agreement is made with sensor chip vendor.

Sign off will also be provided after we know we can publish it.